chore: add /utils dir with own file to each utility
