+++ frontend/src/utils/index.ts
@@ -1,0 +1,8 @@
+export * from './isValidCategory.utility'
+export * from './setUpperCase.utility'
+export * from './localStorage.utility'
+export * from './isArrayOf.utility'
+export * from './isString.utility'
+export * from './isValidNumber.utility'
+export * from './isValidObject.utility'
+export * from './getPriceWithDiscount.utility'
