+++ frontend/src/utils/isString.utility.ts
@@ -1,0 +1,1 @@
+export const isString = (string: unknown): boolean => (string !== null || string !== undefined) && typeof string === 'string'
