+++ frontend/src/utils/isValidObject.utility.ts
@@ -1,0 +1,1 @@
+export const isValidObject = (obj: unknown): boolean => typeof obj === 'object' && obj !== null && Object.keys(obj).length > 1
