+++ frontend/src/utils/setUpperCase.utility.ts
@@ -1,0 +1,1 @@
+export const setUpperCase = (name: string): string => name[0].toLocaleUpperCase() + name.slice(1)
